requirements.txt
aniso8601==7.0.0
asgiref==3.5.2
Django==4.1.1
django-admin-interface==0.20.0
django-colorfield==0.7.2
django-filter==22.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-graphql-jwt==0.3.4
graphene==2.1.9
graphene-django==2.15.0
graphene-gis==0.0.7
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.2.0
promise==2.3
psycopg2==2.9.3
PyJWT==2.5.0
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
